Implemented countof closes #79
